% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma de dos numeros}
\usage{
suma(a, b)
}
\arguments{
\item{a}{primer numero que se desea sumar}

\item{b}{segundo numero que se desea sumar}
}
\value{
retorna un numero
}
\description{
Se usa este espacio para documentar la funcion, primero el titulo,
la segunda linea es la descripcion de lo que hace la funcion y
la tercera linea seran los detalles.
}
\details{
Esta funcion suma dos nu\'meros
}
