% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma de dos nu\'meros}
\usage{
suma(a, b)
}
\arguments{
\item{a}{primer numero que se desea sumar}

\item{b}{segundo numero que se desea sumar}
}
\value{
retorna un numero
}
\description{
Se usa este espacio para documentar la funcion, primero el título,
la segunda línea es la descripción de lo que hace la función y
la tercera línea serán los detalles.
}
\details{
Esta funcion suma dos numeros
}
